backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.20.0  # ASGI server to run FastAPI
sqlalchemy>=2.0.0
pydantic>=1.10.0,<2.0.0 # For data validation (FastAPI depends on v1 for now)
requests>=2.28.0       # For making HTTP requests (to Weatherstack)
python-dotenv>=0.21.0  # To load environment variables (like API keys)
websockets>=10.0       # Underlying library often used by FastAPI WebSockets
gunicorn
psycopg2-binary
opencv-python